+++ src/@locale/app/lab/index.js
@@ -5,6 +5,7 @@ import metrics        from './metrics'
 import scheduler      from './scheduler'
 import schedulerModel from './schedulerModel'
 import schedulerMonth from './schedulerMonth'
+import schedulerWeek  from './schedulerWeek'
 
 const lab =
 {
@@ -15,6 +16,7 @@ const lab =
     scheduler      ,
     schedulerModel ,
     schedulerMonth ,
+    schedulerWeek ,
 };
 
 export default lab ;